added support to get addresses by long lat and distance

+++ migrations/20190724054835-addresses-table.js
@@ -7,8 +7,8 @@ exports.up = (db, callback) => {
     region: { type: `string` },
     zip_code: { type: `string` },
     country: { type: `string` },
-    longitude: { type: `decimal` },
-    latitude: { type: `decimal` },
+    longitude: { type: `float8` },
+    latitude: { type: `float8` },
     timezone: { type: `string` }
   }, callback)
 };
